Actually stop coverity moaning about deliberate assignments.
Coverity complains when assignments are made in the condition for any
function called 'assert', on the assumption that assert will be optimised
out of non-debug builds, thus removing the assignment. Libec uses its own
assert() definition that is always present, even in production, but handles
debug output differently.

As Coverity has no modelling option to suppress the warnings for assignments
in an assert statement, the previous commit renamed assert() to ec_assert()
in an attempt to shut it up. Unfortunately, this didn't solve the issue, as
coverity seems to treat any 'assert-like' function name the same way.

This commit renames it again, to ec_abort(). A test build with these changes
was uploaded to coverity, and confirmed to behave correctly - coverity no
longer complains about the assignment.