Fix cert test script
Certificate was created, but not assigned to the 'c' variable. Due
to (I assume) the new allocation being in the same memory space the
tests were still passing, as there was a valid certificate there with
the correct parameters set, but c was not actually a valid pointer at
the time - it was only by coincidence that it happened to still work.

Assigned the result of ec_create_cert() to c, as was intended all along.