Clamp records added to a NOSIGN section
If records are added to a NOSIGN section, then they must also have
the NOSIGN flag set. This is because signed records in a NOSIGN
section could result in those records changing section if the
section header is removed, even after the certificate is signed, and
the record would now be considered part of the preceding section). A
certificate with such a 'moved' record would still be considered
valid, which therefore makes it an unwanted attack vector. This
commit removes the vector by automatically setting the NOSIGN flag on
any record added to a NOSIGN section.